Create pf.conf
PF Firewall rule, currently set up as such that it logs and pushes the packets to special device in /dev/pflog0. The requirement to run this PF is to init the rc vars in /etc/rc.conf. This firewall rule does NAT and redirection for the jails to receive proper networking from the host.